Fix randVal=0 UB and powers-of-10 bug in SDtest(); SD_CS type
SDtest(): random(30557) could return 0, making log10(0) = -inf and the
subsequent ceil/cast to int undefined behaviour. Also, for any power of
10 (e.g. 10, 100, 1000) ceil(log10(n)) equals the digit count rather
than digit_count-1, so randLength was one short and the last digit was
not verified. Fix both by starting the range at 1 and computing the
length directly from strlen(randDigits)+1 (the +1 accounts for the \r
that println appends; the loop's existing randLength-1 bound strips it).

Margay.h: SD_CS was declared int while every other pin member is
uint8_t; align the type for consistency.

+++ src/Margay.h
@@ -119,7 +119,7 @@ class Margay
     void powerAux(bool state);
 
     //Pin definitions
-    int SD_CS = 4;
+    uint8_t SD_CS = 4;
     uint8_t AuxLED = 20;
     uint8_t RedLED = 13;
     uint8_t GreenLED = 15;
